requirements.txt
llama-cpp-python
pandas
numpy
psutil